pr_optimizer/pyproject.toml
[project]
name = "codejangler"
version = "0.1.0"
description = "AI-powered tool for optimizing and splitting large pull requests"
authors = [{ name = "CodeJangler Team", email = "team@codejangler.dev" }]
requires-python = ">=3.10,<=3.13"
readme = "README.md"
license = { file = "LICENSE" }
keywords = ["git", "github", "pull-request", "ai", "code-review"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Version Control :: Git",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "crewai[tools]>=0.86.0,<1.0.0",
    "click>=8.1.7",
    "gitpython>=3.1.40",
    "pygithub>=2.1.1",
    "pydantic>=2.4.2",
    "rich>=13.5.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "black>=24.1.0",
    "isort>=5.13.0",
    "mypy>=1.8.0",
    "ruff>=0.2.0",
]

[project.urls]
Homepage = "https://github.com/codejangler/codejangler"
Documentation = "https://docs.codejangler.dev"
Repository = "https://github.com/codejangler/codejangler.git"
Issues = "https://github.com/codejangler/codejangler/issues"

[project.scripts]
codejangler = "pr_optimizer.main:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 88
target-version = ["py310", "py311", "py312"]

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-ra -q --cov=pr_optimizer"
testpaths = ["tests"]

[tool.ruff]
line-length = 88
target-version = "py310"
select = ["E", "F", "B", "I"]

[tool.hatch.build.targets.wheel]
packages = ["src/pr_optimizer"]
